requirements.txt
streamlit
pandas
pymongo
requests
duckduckgo-search
wikipedia
google-generativeai
lxml